Jump parameter sets if long periods of silence
If packets are missed, the sender and receiver packet counters can
become desynchronized. If this happens, and the sender moves on to a new
parameter set that is incompatible, the receiver will no longer be able
to receive any packets.

One way to help this would be to automatically resync the packet
counters with every received packet (which I plan to implement). But
this still fails in the case of the last packet of a parameter set being
missed.

To account for this, I have added a backstop time of 10 seconds of no
packet reception, after which the receiver will assume it has gotten
behind the sender and it will automatically move to the next parameter
set in an attempt to catch up.